riscv/signal: add CHERI-variant of rt_sigreturn
Syscalls return to the label ret_from_syscall in entry.S, where their
return value is stored on the stack to be later restored into register
a0. This way, user-space has access to the syscall result. Signal
handlers also use syscalls, specifically the syscall rt_sigreturn. It is
used to complete the execution of a signal handler. In this case, we do
not want to overwrite the value of a0 from the task, which has been
interrupted. rt_sigreturn therefore returns the stored value of register
a0 to ensure the correct value is written to the stack in entry.S.

The return type of syscalls defined using SYSCALL_DEFINEx is long. On
CHERI systems, the tag of capabilities stored in ca0 / a0 will therefore
be lost by rt_sigreturn. Define the syscall using SYSCALL_CHERI_DEFINE0
to allow it to return capabilities.